Add a helper function file and add mapCards to it
