+++ src/Components/EpisodesContainer/AllEpisodesContainer.js
@@ -1,25 +1,6 @@
-import EpisodeCard from "../EpisodeCard/EpisodeCard"
+import { mapCards } from "../../helpers"
 
 const AllEpisodesContainer = ({allEpisodes}) => {
-
-  const mapCards = (episodes) => {
-    console.log('here',episodes)
-    return episodes.map(episode => {
-      return (
-        <EpisodeCard 
-          key={episode.id}
-          id={episode.id}
-          season={episode.season}
-          episode={episode.episode}
-          name={episode.name}
-          rating={episode.rating}
-          description={episode.description}
-          airDate={episode.airDate}
-          thumbnailUrl={episode.thumbnailUrl}
-        />
-      )
-    })
-  } 
 
   const allEpisodeCards = mapCards(allEpisodes)
